CHROMIUM: dark resume: move to general platform api
This changes dark resume so that more control is given to the architecture
specific code to check what woke the system up. A callback that is registered
that checks if any of the devices in the source list woke the system. This
function can call into platform specific code to figure out what resumed the
system.

BUG=chromium:223578
TEST=manual
	Tested on parrot. Run the script activate_short_dark_resume and close the
lid. Check that the system resumes and then suspends again when plugged in.
Check that the system shuts down when unplugged after a few suspend resume
cycles.

Change-Id: If1afd3e6a333494b304feacfed879d85e7cfeb3b
Signed-off-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/58974