CHROMIUM: spi: Add flag to specify CS_SW Control
This patch adds flags to the spi_transfer struct which allow the
spi client to specify that they are going to control the value
of CS. This can be used to control which spi transfers end up
in a single transaction (between CS dropping and rising again).

TEST=Verified that flags were able to be set in client and honored
by the subsequent spi controller driver.
BUG=None

Change-Id: I5862d93bc8aecfa34f5321c274baf6b866177390
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/65879
Reviewed-by: Olof Johansson <olofj@chromium.org>
Commit-Queue: Olof Johansson <olofj@chromium.org>