UPSTREAM: perf: Disable monitoring on setuid processes for regular users
There was a a bug in setup_new_exec(), whereby
the test to disabled perf monitoring was not
correct because the new credentials for the
process were not yet committed and therefore
the get_dumpable() test was never firing.

The patch fixes the problem by moving the
perf_event test until after the credentials
are committed.

Signed-off-by: Stephane Eranian <eranian@google.com>
Tested-by: Jiri Olsa <jolsa@redhat.com>
Acked-by: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: <stable@kernel.org>
Signed-off-by: Ingo Molnar <mingo@kernel.org>

BUG=chromium:256280
TEST=manual, PoC can no longer watch setuid processes

(cherry picked from commit 2976b10f05bd7f6dab9f9e7524451ddfed656a89)
Change-Id: I0f79a3e11977849422a785325a598a76e70b30c1
Signed-off-by: Kees Cook <keescook@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60660
Reviewed-by: Aaron Durbin <adurbin@chromium.org>