ARM: GIC: interface to send a SGI directly
The regular gic_raise_softirq() takes as input a CPU mask which is not
adequate when we need to send an IPI to a CPU which is not represented
in the kernel to GIC mapping.  That is the case with the b.L switcher
when GIC migration to the inbound CPU has not yet occurred.

Signed-off-by: Nicolas Pitre <nico@linaro.org>
(cherry-picked from commit 1c332d4f223066932035578f22877d269443394e)
(cherry-picked from: git://git.linaro.org/people/nico/linux.git)

BUG=chrome-os-partner:19820
TEST=Booted peach

Change-Id: Ib8d37ae97a4cd52b104b363f5e7c02df6b96aacb
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
[abrestic: fixed style errors]
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59037