ARM: Fix dtb dependency to use order-only prerequisites
The %.dtb dependency is specified to depend on the PHONY "scripts".
That means that it'll build every time even if the underlying dtb file
hasn't been touched.  Use an order-only prerequisites to fix this.
Also mark "dtbs" as PHONY for correctness.

This was broken in (70b0476 ARM: 7513/1: Make sure dtc is built before
running it).

BUG=chromium:253588
TEST=manually 'make exynos5250-snow.dtb' and see:
  make[2]: `arch/arm/boot/exynos5250-snow.dtb' is up to date.
TEST=touch 'arch/arm/boot/dts/exynos5250-snow.dts' and repeat.  See:
  DTC     arch/arm/boot/exynos5250-snow.dtb
TEST=manually 'make dtbs' and see that it still works.

This has been posted upstream:
  https://patchwork.kernel.org/patch/2773591/

Change-Id: I4b0fe2c7551c5774ba348c53fd437c978ed6cfcf
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59822
Reviewed-by: Mike Frysinger <vapier@chromium.org>