ARM: Avoid extra calls to the C compiler
Starting up the C compiler can be a slow operation on some systems.
Though these calls don't individually take a lot of time, they add up.
Rearrange the ARM Makefile a bit to avoid extra calls to the compiler
when they can be easily avoided.

When running with the Chrome OS ARM cross compiler
"armv7a-cros-linux-gnueabi-", this shaved .55 seconds (from 5.31
seconds to 4.76 seconds) off an incremental build of the kernel:
  time make -j32 ARCH=arm CROSS_COMPILE=armv7a-cros-linux-gnueabi-

Thanks to Mike Frysinger for the clean trick to make this work.

BUG=chromium:253588
TEST=Build incremental (no-op) build in ChromeOS before and after.
With this change alone we save 4.25 seconds (from 44.141 seconds to
39.891 seconds).
TEST=Build and boot; run trybots.

This patch has been posted upstream:
  https://patchwork.kernel.org/patch/2777841/

Change-Id: I79b32012abd3018578f01cc5841d1e7b1e962463
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/59879