kbuild: Avoid race generating / using MODLIB
The $(MODLIB) variable is recursively expanded and is:
  $(INSTALL_MOD_PATH)/lib/modules/$(KERNELRELEASE)

The $(KERNELRELEASE) variable is also recursively expanded:
  $(shell cat include/config/kernel.release 2> /dev/null)

The file being read is a generated by the Makefile itself.  When the
file is generated it is removed and then recreated.  This leaves a
period of time when the file doesn't exist.  If we happen to use
$(MODLIB) during this time then we'll end up with a bad value.

This is fairly easy for me to reproduce by:
* Add a "sleep 2" between the 'rm -f $@' and the 'echo' in the rule
  to build 'include/config/kernel.release'
* make -j32 ARCH=arm CROSS_COMPILE=armv7a-cros-linux-gnueabi- \
  zImage dtbs modules
* make -j32 ARCH=arm CROSS_COMPILE=armv7a-cros-linux-gnueabi- \
  INSTALL_PATH=foo INSTALL_MOD_PATH=foomod \
  install firmware_install modules_install

Without trying to rethink the whole KERNELRELEASE variable, the simple
fix is to just add a dependency.

This is posted upstream at:
  https://patchwork.kernel.org/patch/2794701/

BUG=chromium:259100
TEST=Add "sleep" as per description above and see that build completes.

Change-Id: I72b862df0d6edc745e535febb460a063901c105d
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61495