CHROMIUM: arm: exynos: Do not disable CHIPID clock
EVT1 iROM code checks the PKG_ID (0x1000_0004) initially and then follows
different sequences based on whether it is an EVT0/EVT1 (primarily the
LOWPOWER_EN flag check). The ASV code disables the CHIPID block clock and
thus also gates the PKG_ID SFR at kernel boot-up.
When core switching occurs, the inbound CPU checks the PKG_ID in the iROM
and takes the wrong path (EVT0 one) due to an incorrect value. This causes
us not to enter low_power_start and switching fails. Fix this by removing
the clock handling code in ASV and adding a CLK_IGNORE_UNUSED flag for
CHIPID so that it is not gated by CCF.

NOTE: This change relies on the assumption that we will not gate CHIPID
in u-boot.

BUG=chrome-os-partner:19820
TEST=Boot on peach

Check the status of the CHIPID clock before and after this change.
BEFORE:
armio 0x10020960
10020960:fffffffe
AFTER:
armio 0x10020960
10020960:ffffffff

Confirm that ASV debug prints at bootup before and after the patch are
the same.
BEFORE:
[    0.607683] Exynos5420 ASV : Lot ID is NZ262[Non Special]
[    0.613138] EXYNOS5420 ASV : NZ262 IDS : 44 HPM : 85
[    0.618173] VDD_ARM ASV group is 7
[    0.621638] VDD_KFC ASV group is 7
AFTER:
[    0.606687] Exynos5420 ASV : Lot ID is NZ262[Non Special]
[    0.612148] EXYNOS5420 ASV : NZ262 IDS : 44 HPM : 85
[    0.617180] VDD_ARM ASV group is 7
[    0.620648] VDD_KFC ASV group is 7

With the on-demand governer check the status of the cores.
localhost ~ # ./prime 1 500000 2000000 1 &
[1] 3976
Starting PID [ 3976 ] on CPUID [ 0 ]
PID [3976]: Compute prime numbers between 500000 and 2000000
localhost ~ # cat /dev/bL_status
		0 1 2 3 L2 CCI
	[A15]	0 0 1 0  1  1
	[A7]	1 1 0 1  1  1

Change-Id: Ib50069fde7a798ebbe38516e2de888e331bb8a7c
Signed-off-by: Thomas Abraham <thomas.ab@samsung.com>
Signed-off-by: Abhilash Kesavan <a.kesavan@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/60654
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Tested-by: Alim Akhtar <alim.akhtar@samsung.com>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>