arm: exynos5420: clear and restore IRAM cpu state register
The cpu state register address is common for all CPUs which tells the
irom code to jump to lower power start if the flag is set. When CPU0
suspends and resumes, it must not go to low power start. So clear the
low power flag before suspending. Now, CPU0 will resume normally and
set the cpu state register with low power flag in U-Boot so that the
secondary cpus boot normally.

We also need to restore this register in case suspend fails for some
reason. In this case, core0 will not go to U-Boot and will try to bring
up the secondary cpus which will fail to boot. So restore the low power
flag register in cpu resume.

BUG=chrome-os-partner:20523
TEST=Boot on peach (evt1) and run S2R.
	suspend_stress_test --count=3000 --suspend_min=12 --backup_rtc

Change-Id: Ief040e1b6d291bf3f7eee3c7eb9f37ef02ccd762
Signed-off-by: Prathyush K <prathyush.k@samsung.com>
Signed-off-by: Alim Akthar <alim.akthar@samsung.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/60867
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>