ARM: dts: Add SDR50 mode for WiFi
We now have code to properly handle the SDIO 3.0 negotiation, so we
can claim UHS on the WiFi slot.  Since we're only driving the WiFi
card at 400MHz input clock (and we know that the WiFi chip only
supports SDR50) we set SDR50 as the max.

BUG=chrome-os-partner:20212
TEST=With other patches in series, see that WiFi shows up as SDR50

Change-Id: Iaa4b12e6f17fc42141446b23248c471799458954
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/61945
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Tested-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Sonny Rao <sonnyrao@chromium.org>