CHROMIUM: dt: Add USB VBUS GPIOs to Peach-Pit device tree
Add device tree entries for the GPIOs that control USB port power on
Pit. Required to get USB working on reworked boards, and may become the
default on future ones. Also allows S3 charging if it is enabled in the
kernel.

BUG=chrome-os-partner:18005
TEST=Get a Pit reworked to use GPIOs instead of the xHC output lines for
USB VBUS control, then test USB on it (especially across S2R).
Cherry-pick CL:58424 on top of this and confirm that you can charge
smartphones in suspend.

Change-Id: I06408798bf6f6a14922026e28e137d240a4758a6
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60889
Reviewed-by: Doug Anderson <dianders@chromium.org>
Tested-by: Doug Anderson <dianders@chromium.org>