ARM: dts: Switch eMMC I/O lines off in sleep
We are turning the 2.85V line (BUCK8) off in sleep.  We should be
turning the I/O lines off too.  Note that without a rework we will end
up pulling the CMD line up while in sleep, but that shouldn't be a big
deal.

BUG=None
TEST=suspend / resume works
TEST=Probe lines in sleep and see that voltage is now off.

Change-Id: Iafa057b008bdd602fe39cb094145d69e1ad9e5a4
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/60885
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>