mmc: dw_mmc: Add suspend_noirq/resume_noirq callbacks for dw_mmc-pltfm
On some devices (like exynos5420) the dw_mmc controller may be in a
strange state after we wake up from sleep.  Add callbacks to allow for
dealing with these quirks.  We use the "_noirq" versions of the
callbacks since in the case of exynos5420 the strange state caused
interrupts to fire so we need to deal with it while interrupts are
still off.

At the moment this support is only added to dw_mmc-pltfm which calls
straight to the callback, since nobody but exynos needs it.  We can
add some levels of indirection (a call into the generic dw_mmc code)
when someone finds a need.

BUG=chrome-os-partner:20409
TEST=Suspend/resume OK now on rev2.

This has been posted upstream at:
  https://patchwork.kernel.org/patch/2825773/

Change-Id: Ib959439f6b7fa41851a516e2124b7e634c098dc6
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: James Hogan <james.hogan@imgtec.com>
Reviewed-on: https://gerrit.chromium.org/gerrit/61242