mmc: dw_mmc: Invalidate cache of current_speed after suspend/resume
The dw_mmc driver keeps a cache of the current slot->clock in order to
avoid doing a whole lot of work every time set_ios() is called.
However, after suspend/resume the register values are bogus so we need
to ensure that the cached value is invalidated.

In many cases we got by without this since the core mmc code fiddles
with the clock a lot.  If we've got a card present we're probably
running it at something like 50MHz and the core will temporarily
switch us to 400kHz after resume.  One case that didn't work (for me)
is the case of having no card in the slot.  The slot is initted to
400kHz at boot time.  After suspend/resume the slot thinks it's still
at 400kHz (due to the cache) so doesn't adjust timing.  When it tries
to send the command at probe time it just times out and gets left in a
bad state.

Invalidating the current_speed also means that we don't need to call:
  dw_mci_setup_bus(slot, true);
...to force an update of the clock in the case when the slot was left
powered.

BUG=chrome-os-partner:20409
TEST=With no card in slot, do suspend/resume.  Then put a SD card in.

This has been posted upstream at:
  https://patchwork.kernel.org/patch/2825478/

Change-Id: I4f20d6a5999a3917271992cf05a711342f8faf96
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61277
Reviewed-by: Simon Glass <sjg@chromium.org>