UPSTREAM: clk: tegra: Fix "defer application of init table"
The patch titled "clk: tegra: defer application of init table" was
misapplied, and dropped an essential change.  The patch was applied
correctly to linux-next, however.

When this patch was merged into the NVIDIA topic branches, it was split
into 2 parts (to avoid merge conflicts). However, when this patch was
merged, only 1/2 of it was merged. This is the remaining half.

Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Andrew Chew <achew@nvidia.com>
Change-Id: Ifbf4929ab33a72be45ef6efae8bf4c2aef12e0d3
Reviewed-on: https://gerrit.chromium.org/gerrit/61776
Commit-Queue: Olof Johansson <olofj@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>