ARM: dts: add max77802 BUCK ramp delays
With the default ramp delay of 100.0mV/us, the voltage undershoot is
too high (about 80mV) which is not giving enough of a voltage margin
when the frequency is changing.  By setting the ramp delay to 12.5mV/us,
the voltage undershoot is reduced significatly to about 30mV, which gives
a good enough voltage margin when the frequency changes.

BUG=chrome-os-partner:20523,chrome-os-partner:20829
TEST=Boot Pit with DVFS enabled; observe that suspend/resume is more
reliable (over 4000 cycles without failure).

Change-Id: I0d1f1287d40c91df619f36d9858a8f79e915f532
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61887
Reviewed-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Simon Glass <sjg@chromium.org>