arm: exynos: disable all interrupts on Exynos5420 before suspend
Disable all interrupts from the GIC before entering suspend on
Exynos5420 as is done on Exynos5250.  If interrupts are enabled, we
may receive an interrupt after entering WFI but before the PMU has
suspended the system, causing suspend to fail.

BUG=chrome-os-partner:20523
TEST=Run suspend_stress_test on Pit and observe that entering suspend
no longer occasionally fails with the "Failed to suspend the system"
error in exynos_cpu_suspend().

Change-Id: Ia2d963191f4e5485beb295b28f128f2ce256f987
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/61948
Reviewed-by: Simon Glass <sjg@chromium.org>