mmc: dw_mmc: recover from Start Bit Error
During a 4-bit or 8-bit read-data transfer, if data line does
not have a start bit, the data path signals a data start bit error to
the BIU and waits for a data timeout, after which it signals that the
data transfer is done. This patch sets the Data transfer done signals
once SBE interrupt reported.
Also this patch reset the DMA once the error is generated, to bring the DMA
to the default state.

This patch also makes the error reporting more verbose.

BUG=chrome-os-partner:20866
TEST=build and booted on peach, with HS200 mode enabled, not seeing the
     board hangs, also HS200 mode recover and function normally.
     Run multiple reboot and no hang seen.
     I can see the Start Bit Error oocured and system is able to recover
     from it.
TEST=manual, can see now correct error being reported
     by kernel.
     /wo this patch, even for Start Bit Error, log was showing as a DATA
     FIFO error
     /w this CL, below error reported
     [    4.700880] mmc_host mmc0: Bus speed (slot 0) = 200000000Hz
     (slot req 200000000Hz, actual 200000000HZ div = 0)
     [    4.710305] dwmmc_exynos 12200000.dwmmc0: Start bit error
     (status=00002000)

Change-Id: Ic11849eb98750c4a20e68c5bb4b3cda648c0d604
Signed-off-by: Alim Akhtar <alim.akhtar@samsung.com>
Signed-off-by: Doug Anderson <dianders@chromium.org>
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62318