CHROMIUM: spring: dts: disable camera FET during suspend
The USB camera firmware is not always suspending properly due to the
glitches on the XHCI controller line when entering suspend,
so we cut the power rail to the camera to avoid leaking power.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BUG=chrome-os-partner:19915
TEST=On Spring, suspend the machine and verify that the FET5 is off by
typing "pmu" on the EC console and seeing register 0x14 power good bit
unset, then resume and check that the camera is working.

Change-Id: I39e5e659ace0f8c73fce62a9df3b0011de42020f
Reviewed-on: https://gerrit.chromium.org/gerrit/62948
Reviewed-by: Todd Broch <tbroch@chromium.org>
Commit-Queue: Vincent Palatin <vpalatin@chromium.org>
Tested-by: Vincent Palatin <vpalatin@chromium.org>