mcpm: exynos: disable own CCI port if other cluster is powered off
There is a brief race when the A7 cluster is powering on the A15
cluster where the A7's CCI port is enabled and the A15's have just
powered up, but have yet to disable L2 auto clock-gating.  Snoops
from the A7 cluster to the gated A15 cluster's L2 cache will then
cause the system to hang.  To work around this, disable the CCI
ports of both clusters when either of the clusters is completely
powered off.  The CCI port of the powered on cluster will then
be re-enabled once the powered-down cluster is switched on.

BUG=chrome-os-partner:21189
TEST=Boot on peach, run Video and see IKS working. Verify that when a
cluster is down, both the snoops are off using bL_status.
	cat /dev/bL_status
		0 1 2 3 L2 COMMON CCI
	[A15]	1 1 1 1  1    1    0
	[A7]	0 0 0 0  0    0    0

Change-Id: Ie64d6a18951453c6b130ec836217688b095185de
Signed-off-by: Thomas Abraham <thomas.abraham@linaro.org>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62802
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Reviewed-by: Thomas Abraham <thomas.ab@samsung.com>