CHROMIUM: dt: Add ramoops entry in Peach-Pit device tree
This patch adds the device tree entry necessary to reserve a memory
region for ramoops to the Peach-Pit device tree. It just copies the same
region that was used on Snow (cros5250-common), since I was told that
the region is boring old (and in no way special) DRAM and it's position
doesn't really matter.

BUG=chrome-os-partner:20869
TEST=Run 'echo c > /proc/sysrq-trigger' on Pit, make sure you get a
kcrash in /var/spool/crash after the reboot.

Change-Id: I292da60db2f01a7821591cc7a7c2b99a0b7a7cbb
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62742
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>
Commit-Queue: Doug Anderson <dianders@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>