ARM: dts: Add tps65090 charger to peach-pit
We've added polling support for the charger in the case that we don't
have a nice clean IRQ.  We'll use that for pit.  We sort of do have an
IRQ (the tpschrome IRQ is routed to the AP), but the EC is in charge
of managing that IRQ and trying to hack up TPSChrome to look at the
IRQs but not touch them is quite ugly.

BUG=chrome-os-partner:21151
TEST=Unplug power and see charge status change.

Change-Id: I0a2ebece37e9985f99ff95dfcfefc8aeafaf6525
Signed-off-by: Doug Anderson <dianders@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62817