HACK: mcpm: exynos: Always keep A7 L2 COMMON on
Always leaving A7 L2 COMMON on, even if all A7 cores are power-gated,
seems to make IKS more stable.  This should have a minimal power impact
since: 1) the A7 cluster is low-power to begin with, and 2) the only
situation in which this makes a difference is when all A15s are powered
up.

This may be related to ARM errata 802022 which only occurs when the A7
cluster is completely powered on or off.  A workaround for this errata,
however, requires a lot of coordination between the A7s when powering
on or off.

BUG=chrome-os-partner:21189
TEST=IKS appears to be much more stable.  mem-stress-switcher completes
multiple iterations without hanging.  Previously it would hang within
5 minutes into the first iteration.

Change-Id: I5cd1a38f8cdae0f78d4531faa13c0e9c32232ffb
Signed-off-by: Arjun.K.V <arjun.kv@samsung.com>
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/62803
Reviewed-by: Doug Anderson <dianders@chromium.org>