CHROMIUM: pci: default d3cold_allowed to false
We are not quite ready to enable d3cold just yet. Default
d3cold_allowed to false for now.  We will want to revisit
this in the not too distant a future, though.

BUG=chrome-os-partner:21524
TEST=run the command below and check to make
sure that it returns 0 for all devices:
for i in /sys/devices/pci*/00*; do cat $i/d3cold_allowed; done

Change-Id: I9e26fa952c6b481d929374e856c5b1df7a6b3871
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/63981
Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>