drm/exynos: Fix interlaced modes
A previous patch removed a call to drm_mode_set_crtcinfo in the hdmi
driver with flag zero, this was overwriting drm subsystem's default of
halving vertical timings when interlaced. To accommodate the doubled
vertical values, the hdmi driver was dividing vertical values manually.
This patch removes that division and instead uses the values directly.

BUG=chromium:253382
TEST=Tested on snow with both interlaced and progressive modes

Change-Id: Id1cdb36b17627ad7cd84204e8fd2ad1244929cbb
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/64833