drm/exynos: Add 800x600 and 1440x900 modes to HDMI
This patch adds 800x600 and 1440x900 modes. The 1440x900 mode is
supported by faking the output image size to be 224 horizontal lines
bigger, and to blank an extra 224 lines.

The 800x600 mode uses this offset as well, but it also activates the
stereoscopic function of the mixer such that it scans 832 horizontally
instead of 1632, then the blank is increased by 32 lines like above to
get 800 columns.

BUG=chromium:221411, chrome-os-partner:14238
TEST=Tested on snow with 800x600 by hand

Change-Id: I051f7c24143acb63f477319616c61966d54be173
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Signed-off-by: Shirish S <s.shirish@samsung.com>
Reviewed-on: https://chromium-review.googlesource.com/64834